Move templates to templates/package.dart
